automatically backed up by learn

+++ lib/create.sql
@@ -2,7 +2,7 @@ CREATE TABLE projects (
   id INTEGER PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal FLOAT,
+  funding_goal INTEGER,
   start_date_field INTEGER,
   end_date_field INTEGER
 );
